requirements.txt
blinker==1.8.0
click==8.1.7
colorama==0.4.6
flask==3.0.3
flask-mail==0.9.1
flask-sqlalchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5
python-dotenv==1.0.1
sqlalchemy==2.0.29
typing-extensions==4.11.0
werkzeug==3.0.2
